backend/requirements.txt
absl-py==1.4.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==3.7.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
astunparse==1.6.3
async-lru==2.0.4
attrs==23.1.0
axios==0.4.0
Babel==2.12.1
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.0.0
blinker==1.6.2
blis==0.7.11
cachetools==5.3.1
catalogue==2.0.10
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.3
cloudpathlib==0.18.1
cmake==3.29.5.1
colorama==0.4.6
comm==0.1.4
confection==0.1.5
contourpy==1.1.0
cycler==0.11.0
cymem==2.0.8
dataclasses-json==0.6.7
debugpy==1.6.7.post1
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.6
distro==1.9.0
docx2txt==0.9
executing==1.2.0
faiss-cpu==1.13.2
fastjsonschema==2.18.0
filelock==3.12.2
Flask==2.3.2
Flask-Cors==4.0.1
flatbuffers==24.3.25
fonttools==4.41.1
fqdn==1.5.1
frozenlist==1.8.0
fsspec==2025.9.0
gast==0.4.0
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
greenlet==3.0.3
groq==0.8.0
grpcio==1.56.0
h11==0.14.0
h5py==3.11.0
hf-xet==1.2.0
httpcore==1.0.5
httpx==0.27.0
httpx-sse==0.4.3
huggingface_hub==1.4.1
idna==3.4
ipykernel==6.25.1
ipython==8.14.0
ipython-genutils==0.2.0
ipywidgets==8.1.0
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.19.0
Jinja2==3.1.2
jiter==0.13.0
joblib==1.3.2
json5==0.9.14
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.7.0
jupyter-lsp==2.2.0
jupyter_client==8.3.0
jupyter_core==5.3.1
jupyter_server==2.7.0
jupyter_server_terminals==0.4.4
jupyterlab==4.0.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.8
jupyterlab_server==2.24.0
keras==3.4.1
kiwisolver==1.4.4
langchain==1.2.10
langchain-classic==1.0.1
langchain-community==0.4.1
langchain-core==1.2.11
langchain-text-splitters==1.1.0
langcodes==3.4.0
langgraph==1.0.8
langgraph-checkpoint==4.0.0
langgraph-prebuilt==1.0.7
langgraph-sdk==0.3.5
langsmith==0.7.1
language_data==1.2.0
libclang==16.0.6
lxml==5.3.0
mahotas==1.4.13
marisa-trie==1.2.0
Markdown==3.4.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.26.2
matplotlib==3.7.2
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.1
ml-dtypes==0.4.0
mpmath==1.3.0
multidict==6.7.1
murmurhash==1.0.10
mypy_extensions==1.1.0
namex==0.0.8
nbclient==0.8.0
nbconvert==7.7.3
nbformat==5.9.2
nest-asyncio==1.5.7
networkx==3.4.2
nltk==3.8.1
notebook==7.0.2
notebook_shim==0.2.3
numpy==1.26.4
oauthlib==3.2.2
openai==2.20.0
opencv-python==4.8.0.74
opt-einsum==3.3.0
optree==0.12.1
orjson==3.11.7
ormsgpack==1.12.2
overrides==7.4.0
packaging==26.0
pandas==2.0.3
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
Pillow==10.0.0
platformdirs==3.8.0
preshed==3.0.9
prometheus-client==0.17.1
prompt-toolkit==3.0.39
propcache==0.4.1
protobuf==4.23.4
psutil==5.9.5
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyAudio==0.2.13
pycparser==2.21
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
Pygments==2.16.1
pyparsing==3.0.9
pypdf==6.7.0
python-dateutil==2.8.2
python-dotenv==1.1.0
python-json-logger==2.0.7
pytz==2023.3
pywin32==306
pywinpty==2.0.11
PyYAML==6.0.1
pyzmq==25.1.1
qtconsole==5.4.3
QtPy==2.3.1
referencing==0.30.2
regex==2024.7.24
requests==2.32.5
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rpds-py==0.9.2
rsa==4.9
safetensors==0.7.0
scikit-learn==1.6.1
scipy==1.11.1
Send2Trash==1.8.2
sentence-transformers==5.2.2
shellingham==1.5.4
six==1.16.0
smart-open==7.0.4
sniffio==1.3.0
soupsieve==2.4.1
spacy==3.7.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SpeechRecognition==3.10.0
SQLAlchemy==2.0.31
srsly==2.4.8
stack-data==0.6.2
sympy==1.14.0
tenacity==9.1.4
tensorboard==2.17.0
tensorboard-data-server==0.7.1
tensorflow==2.17.0
tensorflow-estimator==2.13.0
tensorflow-intel==2.17.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
terminado==0.17.1
thinc==8.2.5
threadpoolctl==3.2.0
timm==1.0.24
tinycss2==1.2.1
tokenizers==0.22.2
torch==2.9.0
torchvision==0.24.0
tornado==6.3.2
tqdm==4.66.4
traitlets==5.9.0
transformers==5.1.0
typer==0.12.3
typer-slim==0.21.2
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2023.3
uri-template==1.3.0
urllib3==1.26.16
uuid_utils==0.14.0
virtualenv==20.23.1
wasabi==1.1.3
wcwidth==0.2.6
weasel==0.4.1
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.1
Werkzeug==2.3.6
widgetsnbextension==4.0.8
wrapt==1.15.0
xxhash==3.6.0
yarl==1.22.0
zstandard==0.25.0
langchain-huggingface>=0.1.0
